requirements.txt
alembic==1.0.1
blinker==1.4
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-markdown2==0.0.3
Flask-Migrate==2.3.0
Flask-Psycopg2==1.3
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
ItsDangerous==1.0.0
Jinja2==2.10
jinja2-markdown==0.0.3
Mako==1.0.7
Markdown==3.0.1
markdown2==2.3.6
MarkupSafe==1.0
psycopg2==2.7.5
PyJWT==1.4.2
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.12
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
